fix: update search string from query params
